-fixed warnings and errors reported by gcc 2.95 on alpha
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1402 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/domainpolicy/domainpolicy.c
@@ -661,7 +661,7 @@ int dp_can_connect_str(str *domain, int rec_level) {
 	 * wrong kind of terminal
 	 */
 	if ((naptr->flags_len != 1) || (tolower(naptr->flags[0]) != 'u')) {
-	    	LOG(L_ERR,"dp_can_connect_str(): Terminal NAPTR needs flag = 'u' and not '%.*s'.\n", naptr->flags_len, ZSW(naptr->flags));
+	    	LOG(L_ERR,"dp_can_connect_str(): Terminal NAPTR needs flag = 'u' and not '%.*s'.\n", (int)naptr->flags_len, ZSW(naptr->flags));
 		/*
 		 * It's not that clear what we should do now: Ignore this records or regard it as failed.
 		 * We go with "ignore" for now.
